Splitting tests to a new file to avoid file_length
