+++ Tests/SwiftLintFramework/RulesTests.swift
@@ -12,53 +12,6 @@ import XCTest
 // swiftlint:disable:next type_body_length
 class RulesTests: XCTestCase {
 
-    func testAttributes() {
-        // Test with default parameters
-        verifyRule(AttributesRule.description)
-
-        // Test with custom `always_on_same_line`
-        let alwaysOnSameLineDescription = RuleDescription(
-            identifier: "attributes_rule",
-            name: "Attributes",
-            description: "Attributes should be on their own lines in functions and types, " +
-                         "but on the same line as variables and imports",
-            nonTriggeringExamples: [
-                "@objc var x: String",
-                "@objc func foo()",
-                "@nonobjc\n func foo()"
-            ],
-            triggeringExamples: [
-                "@objc\n var x: String",
-                "@objc\n func foo()",
-                "@nonobjc func foo()"
-            ]
-        )
-
-        verifyRule(alwaysOnSameLineDescription,
-                   ruleConfiguration: ["always_on_same_line": ["@objc"]])
-
-        // Test with custom `always_on_line_above`
-        let alwaysOnNewLineDescription = RuleDescription(
-            identifier: "attributes_rule",
-            name: "Attributes",
-            description: "Attributes should be on their own lines in functions and types, " +
-            "but on the same line as variables and imports",
-            nonTriggeringExamples: [
-                "@objc\n var x: String",
-                "@objc\n func foo()",
-                "@nonobjc\n func foo()"
-            ],
-            triggeringExamples: [
-                "@objc var x: String",
-                "@objc func foo()",
-                "@nonobjc func foo()"
-            ]
-        )
-
-        verifyRule(alwaysOnNewLineDescription,
-                   ruleConfiguration: ["always_on_line_above": ["@objc"]])
-    }
-
     func testClosingBrace() {
         verifyRule(ClosingBraceRule.description)
     }
